# docs/api/openapi.yaml
openapi: 3.1.0
paths:
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/AuthTokensDto"
      security:
        - bearer: []
      tags:
        - auth
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/AuthTokensDto"
      security:
        - bearer: []
      tags:
        - auth
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutDto"
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      tags:
        - auth
  /auth/change-password:
    post:
      operationId: AuthController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/AuthTokensDto"
      security:
        - bearer: []
      tags:
        - auth
  /auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/PublicUserDto"
      security:
        - bearer: []
      tags:
        - auth
  /roles:
    get:
      operationId: RolesController_list
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque cursor from the previous page meta
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleDto"
                  meta:
                    type: object
                    required:
                      - cursor
                      - limit
                    properties:
                      cursor:
                        type:
                          - string
                          - "null"
                      limit:
                        type: number
      security:
        - bearer: []
      tags:
        - roles
    post:
      operationId: RolesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/RoleDto"
      security:
        - bearer: []
      tags:
        - roles
  /roles/{id}:
    get:
      operationId: RolesController_get
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/RoleDto"
      security:
        - bearer: []
      tags:
        - roles
    patch:
      operationId: RolesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/RoleDto"
      security:
        - bearer: []
      tags:
        - roles
    delete:
      operationId: RolesController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      tags:
        - roles
  /users:
    get:
      operationId: UsersController_list
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque cursor from the previous page meta
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicUserDto"
                  meta:
                    type: object
                    required:
                      - cursor
                      - limit
                    properties:
                      cursor:
                        type:
                          - string
                          - "null"
                      limit:
                        type: number
      security:
        - bearer: []
      tags:
        - users
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/PublicUserDto"
      security:
        - bearer: []
      tags:
        - users
  /users/{id}:
    get:
      operationId: UsersController_get
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/PublicUserDto"
      security:
        - bearer: []
      tags:
        - users
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/PublicUserDto"
      security:
        - bearer: []
      tags:
        - users
    delete:
      operationId: UsersController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /users/{id}/roles:
    post:
      operationId: UsersController_assignRole
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleDto"
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /users/{id}/roles/{roleId}:
    delete:
      operationId: UsersController_unassignRole
      parameters:
        - name: id
          required: true
          in: path
          description: UUID or erc:<externalReferenceCode>
          schema:
            type: string
        - name: roleId
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      tags:
        - users
  /health:
    get:
      operationId: HealthController_check
      parameters: []
      responses:
        "200":
          description: ""
      tags:
        - Health
info:
  title: Neriva API
  description: Headless-first CMS REST API. Success responses use a { data, meta }
    envelope; errors are RFC 7807 problem+json. URL ids accept a UUID or
    erc:<externalReferenceCode>.
  version: 0.1.0
  contact: {}
tags: []
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    AuthTokensDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        mustChangePassword:
          type: boolean
          description: When true, every request except password change and logout returns
            403
      required:
        - accessToken
        - refreshToken
        - mustChangePassword
    LoginDto:
      type: object
      properties:
        email:
          type: string
          format: email
          example: admin@neriva.com
        password:
          type: string
      required:
        - email
        - password
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    LogoutDto:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    ChangePasswordDto:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword
    PublicUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalReferenceCode:
          type: string
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type:
            - object
            - "null"
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        mustChangePassword:
          type: boolean
        customFields:
          type: object
          additionalProperties: true
      required:
        - id
        - externalReferenceCode
        - tenantId
        - createdAt
        - updatedAt
        - createdBy
        - email
        - displayName
        - mustChangePassword
        - customFields
    PermissionDto:
      type: object
      properties:
        resourceType:
          type: string
          example: page
          description: Resource type, or '*' for all
        action:
          type: string
          example: create
          description: create/read/update/delete/publish, or '*'
      required:
        - resourceType
        - action
    RoleDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalReferenceCode:
          type: string
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type:
            - object
            - "null"
          format: uuid
        name:
          type: string
        description:
          type:
            - object
            - "null"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionDto"
      required:
        - id
        - externalReferenceCode
        - tenantId
        - createdAt
        - updatedAt
        - createdBy
        - name
        - description
        - permissions
    CreateRoleDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        externalReferenceCode:
          type: string
          description: Stable code for idempotent upsert; generated when omitted
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionDto"
      required:
        - name
    UpdateRoleDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        permissions:
          description: Replaces the full permission set
          type: array
          items:
            $ref: "#/components/schemas/PermissionDto"
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
        password:
          type: string
          minLength: 8
          description: Temporary password; a change is forced on first login
        externalReferenceCode:
          type: string
          description: Stable code for idempotent upsert; generated when omitted
        roleIds:
          type: array
          items:
            type: string
            format: uuid
      required:
        - email
        - displayName
        - password
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
        displayName:
          type: string
    AssignRoleDto:
      type: object
      properties:
        roleId:
          type: string
          format: uuid
      required:
        - roleId